+++ recipe/conda_forge_ci_setup/upload_or_check_non_existence.py
@@ -16,17 +16,7 @@ from conda_build.conda_interface import get_index
 import conda_build.api
 import conda_build.config
 
-from .feedstock_outputs import request_copy
-
-
-def split_pkg(pkg):
-    if not pkg.endswith(".tar.bz2"):
-        raise RuntimeError("Can only process packages that end in .tar.bz2")
-    pkg = pkg[:-8]
-    plat, pkg_name = pkg.split(os.path.sep)
-    name_ver, build = pkg_name.rsplit('-', 1)
-    name, ver = name_ver.rsplit('-', 1)
-    return plat, name, ver, build
+from .feedstock_outputs import request_copy, split_pkg
 
 
 @contextlib.contextmanager
